implemented queueing

+++ cwlab/exec/exec.py
@@ -53,7 +53,7 @@ def query_info_from_db(job_id):
                 sleep(retry_delay + retry_delay*random())
     return db_job_id_request
 
-def exec_runs(job_id, run_ids, exec_profile_name, cwl, user_id):
+def exec_runs(job_id, run_ids, exec_profile_name, cwl, user_id=None, max_parrallel_exec_user_def=None):
     
     # check if runs are already running:
     already_running_runs = []
@@ -65,9 +65,15 @@ def exec_runs(job_id, run_ids, exec_profile_name, cwl, user_id):
             run_info =  db_run_id_request.filter(Exec.id==max([r.id for r in db_run_id_request])).first()
             if run_info.time_finished is None:
                 already_running_runs.append(run_id)
-    run_ids = list(set(run_ids) - set(already_running_runs))
+    run_ids = sorted(list(set(run_ids) - set(already_running_runs)))
 
     # create new exec entry in database:
+    if not max_parrallel_exec_user_def is None and \
+        app.config["EXEC_PROFILES"][exec_profile_name]["allow_user_decrease_max_parallel_exec"] and \
+        max_parrallel_exec_user_def < app.config["EXEC_PROFILES"][exec_profile_name]["max_parallel_exec"]:
+        max_parallel_exec = max_parrallel_exec_user_def
+    else:
+        max_parallel_exec = app.config["EXEC_PROFILES"][exec_profile_name]["max_parallel_exec"]
     exec_db_entry = {}
     for run_id in run_ids:
         exec_db_entry[run_id] = Exec(
@@ -84,6 +90,7 @@ def exec_runs(job_id, run_ids, exec_profile_name, cwl, user_id):
             time_started=datetime.now(),
             time_finished=None, #*
             timeout_limit=None, #*
+            max_parallel_exec=max_parallel_exec,
             pid=-1, #*
             user_id=user_id if not user_id is None else None,
             exec_profile=app.config["EXEC_PROFILES"][exec_profile_name],
@@ -91,8 +98,8 @@ def exec_runs(job_id, run_ids, exec_profile_name, cwl, user_id):
         )
         #* will be set by the background process itself
         db.session.add(exec_db_entry[run_id])
-    
     db_commit()
+    
 
     # start the background process:
     # the child process will be detached from the parent
@@ -210,13 +217,12 @@ def terminate_runs(
     for run_id in run_info.keys():
         if isinstance(run_info[run_id]["time_started"], datetime) and \
             not isinstance(run_info[run_id]["time_finished"], datetime):
-            p = Process(run_info[run_id]["pid"])
-            print(run_info[run_id]["pid"])
-            is_killed = kill_proc_tree(run_info[run_id]["pid"])
-            if not is_killed:
-                could_not_be_terminated.append(run_id)
-                continue
-            cleanup_zombie_process(run_info[run_id]["pid"])
+            if run_info[run_id]["pid"] != -1:
+                is_killed = kill_proc_tree(run_info[run_id]["pid"])
+                if not is_killed:
+                    could_not_be_terminated.append(run_id)
+                    continue
+                cleanup_zombie_process(run_info[run_id]["pid"])
             db_run_entry = db_request.filter(Exec.id==run_info[run_id]["db_id"])
             db_run_entry.time_finished = datetime.now()
             db_run_entry.status = "terminated by user"
